requirements.txt
# Core Web Framework
fastapi>=0.115.0        # Latest stable (0.124.0 available but breaking changes)
granian>=1.6.0          # Latest: 2.6.0 (Apr 2025), using stable 1.x
pydantic>=2.10.0        # Latest 2.x stable
uvloop>=0.21.0          # Latest stable

# Data Processing & Analytics  
polars>=1.17.0          # Latest 1.x (released Jul 2024)
duckdb>=1.1.3           # Latest stable (1.4.2 available)
numpy>=2.2.0            # Latest 2.x stable
pyarrow>=18.1.0         # Latest stable

# Serialization & Validation
orjson>=3.10.0          # Latest stable
msgspec>=0.18.6         # Latest stable

# Caching
valkey>=6.0.2           # Valkey client (Redis fork, better performance)

# Database Drivers
asyncpg>=0.31.0         # Fastest async PostgreSQL driver

# Compression & Hashing
zstandard>=0.23.0       # Fast compression with good ratios
xxhash>=3.5.0           # Ultra-fast non-cryptographic hashing

# Rust-powered Tools
minijinja>=2.4.0        # Latest stable
tantivy>=0.22.0         # Latest stable (0.25.1 available)

# OpenTelemetry (latest)
opentelemetry-distro>=0.48b0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
